added ingestion playbook
